Parser: split out tests relating to data sources
