+++ vlog4j-parser/src/test/java/org/semanticweb/vlog4j/syntax/parser/RuleParserTest.java
@@ -9,9 +9,9 @@ package org.semanticweb.vlog4j.syntax.parser;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -21,31 +21,21 @@ package org.semanticweb.vlog4j.syntax.parser;
  */
 import static org.junit.Assert.assertEquals;
 
-import java.io.File;
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.Arrays;
 
 import org.junit.Test;
 import org.semanticweb.vlog4j.core.model.api.Conjunction;
 import org.semanticweb.vlog4j.core.model.api.Constant;
-import org.semanticweb.vlog4j.core.model.api.DataSourceDeclaration;
 import org.semanticweb.vlog4j.core.model.api.Fact;
 import org.semanticweb.vlog4j.core.model.api.Literal;
 import org.semanticweb.vlog4j.core.model.api.PositiveLiteral;
-import org.semanticweb.vlog4j.core.model.api.Predicate;
 import org.semanticweb.vlog4j.core.model.api.PrefixDeclarations;
 import org.semanticweb.vlog4j.core.model.api.Rule;
 import org.semanticweb.vlog4j.core.model.api.Statement;
 import org.semanticweb.vlog4j.core.model.api.Variable;
 import org.semanticweb.vlog4j.core.model.implementation.AbstractConstantImpl;
-import org.semanticweb.vlog4j.core.model.implementation.DataSourceDeclarationImpl;
 import org.semanticweb.vlog4j.core.model.implementation.Expressions;
-import org.semanticweb.vlog4j.core.reasoner.implementation.CsvFileDataSource;
-import org.semanticweb.vlog4j.core.reasoner.implementation.RdfFileDataSource;
-import org.semanticweb.vlog4j.core.reasoner.implementation.SparqlQueryResultDataSource;
 import org.semanticweb.vlog4j.parser.ParsingException;
 import org.semanticweb.vlog4j.parser.RuleParser;
 
@@ -349,49 +339,6 @@ public class RuleParserTest {
 	public void tesLiteralError() throws ParsingException {
 		String input = "<http://example.org/p>(?X,<http://example.org/c)";
 		RuleParser.parseLiteral(input);
-	}
-
-	@Test
-	public void testCsvSource() throws ParsingException, IOException {
-		String input = "@source p(2) : load-csv(\"src/main/data/input/example.csv\") .";
-		ArrayList<Statement> statements = new ArrayList<>(RuleParser.parse(input).getStatements());
-		CsvFileDataSource csvds = new CsvFileDataSource(new File("src/main/data/input/example.csv"));
-		Predicate p = Expressions.makePredicate("p", 2);
-		DataSourceDeclaration d = new DataSourceDeclarationImpl(p, csvds);
-		assertEquals(Arrays.asList(d), statements);
-	}
-
-	@Test
-	public void testRdfSource() throws ParsingException, IOException {
-		String input = "@source p(3) : load-rdf(\"src/main/data/input/example.nt.gz\") .";
-		ArrayList<Statement> statements = new ArrayList<>(RuleParser.parse(input).getStatements());
-		RdfFileDataSource rdfds = new RdfFileDataSource(new File("src/main/data/input/example.nt.gz"));
-		Predicate p = Expressions.makePredicate("p", 3);
-		DataSourceDeclaration d = new DataSourceDeclarationImpl(p, rdfds);
-		assertEquals(Arrays.asList(d), statements);
-	}
-
-	@Test(expected = ParsingException.class)
-	public void testRdfSourceInvalidArity() throws ParsingException, IOException {
-		String input = "@source p(2) : load-rdf(\"src/main/data/input/example.nt.gz\") .";
-		RuleParser.parse(input);
-	}
-
-	@Test
-	public void testSparqlSource() throws ParsingException, MalformedURLException {
-		String input = "@source p(2) : sparql(<https://query.wikidata.org/sparql>,\"disease, doid\",\"?disease wdt:P699 ?doid .\") .";
-		ArrayList<Statement> statements = new ArrayList<>(RuleParser.parse(input).getStatements());
-		SparqlQueryResultDataSource sparqlds = new SparqlQueryResultDataSource(
-				new URL("https://query.wikidata.org/sparql"), "disease, doid", "?disease wdt:P699 ?doid .");
-		Predicate p = Expressions.makePredicate("p", 2);
-		DataSourceDeclaration d = new DataSourceDeclarationImpl(p, sparqlds);
-		assertEquals(Arrays.asList(d), statements);
-	}
-
-	@Test(expected = ParsingException.class)
-	public void testSparqlSourceMalformedUrl() throws ParsingException, MalformedURLException {
-		String input = "@source p(2) : sparql(<not a URL>,\"disease, doid\",\"?disease wdt:P699 ?doid .\") .";
-		RuleParser.parse(input);
 	}
 
 	@Test(expected = ParsingException.class)
